refactor: extract getters and setters methods to their own traits
